src/comisiones/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==1.4.48
pymongo==4.3.3
python-pulsar==3.4.0
python-multipart==0.0.6
pydantic[email]==2.5.0
requests==2.31.0
python-dotenv==1.0.0

alembic==1.12.1

pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

python-dateutil==2.8.2
pytz==2023.3
